+++ MadgwickAHRS.h
@@ -20,13 +20,13 @@
 //----------------------------------------------------------------------------------------------------
 // Variable declaration
 
-extern float madgwick_beta;   // algorithm gain
-extern float q_a[4];  // quaternion of sensor frame relative to auxiliary frame
+extern double madgwick_beta;   // algorithm gain
+extern double q_a[4];  // quaternion of sensor frame relative to auxiliary frame
 
 //---------------------------------------------------------------------------------------------------
 // Function declarations
 
-void MadgwickAHRSupdate(float gx, float gy, float gz, float ax, float ay, float az, float mx, float my, float mz, float deltat);
+void MadgwickAHRSupdate(double gx, double gy, double gz, double ax, double ay, double az, double mx, double my, double mz, double deltat);
 
 #endif
 //=====================================================================================================
